add io module for stdin, stderr, stdout functions
